requirements.txt
pandas==1.4.2
pyodbc==4.0.32
PyYAML==6.0
SQLAlchemy==1.4.32